+++ exonum/src/node/mod.rs
@@ -26,7 +26,7 @@ pub use self::{
 pub mod state;
 
 use failure::Error;
-use futures::{sync::mpsc, Future, Sink};
+use futures::{sync::mpsc, Sink};
 use tokio_core::reactor::Core;
 use tokio_threadpool::Builder as ThreadPoolBuilder;
 use toml::Value;
@@ -53,7 +53,7 @@ use crate::events::{
     error::{into_failure, LogError},
     noise::HandshakeParams,
     HandlerPart, InternalEvent, InternalPart, InternalRequest, NetworkConfiguration, NetworkEvent,
-    NetworkPart, NetworkRequest, SyncSender, TimeoutRequest,
+    NetworkPart, NetworkRequest, SyncSender, TimeoutRequest, UnboundedSyncSender,
 };
 use crate::helpers::{
     config::ConfigManager,
@@ -113,7 +113,7 @@ pub trait SystemStateProvider: ::std::fmt::Debug + Send + 'static {
 
 /// Transactions sender.
 #[derive(Clone)]
-pub struct ApiSender(pub mpsc::Sender<ExternalMessage>);
+pub struct ApiSender(pub mpsc::UnboundedSender<ExternalMessage>);
 
 /// Handler that that performs consensus algorithm.
 pub struct NodeHandler {
@@ -337,7 +337,7 @@ pub struct NodeSender {
     /// Network requests sender.
     pub network_requests: SyncSender<NetworkRequest>,
     /// Api requests sender.
-    pub api_requests: SyncSender<ExternalMessage>,
+    pub api_requests: UnboundedSyncSender<ExternalMessage>,
 }
 
 /// Node role.
@@ -772,7 +772,7 @@ impl fmt::Debug for NodeHandler {
 
 impl ApiSender {
     /// Creates new `ApiSender` with given channel.
-    pub fn new(inner: mpsc::Sender<ExternalMessage>) -> Self {
+    pub fn new(inner: mpsc::UnboundedSender<ExternalMessage>) -> Self {
         ApiSender(inner)
     }
 
@@ -786,11 +786,11 @@ impl ApiSender {
     pub fn send_external_message(&self, message: ExternalMessage) -> Result<(), Error> {
         self.0
             .clone()
-            .send(message)
-            .wait()
+            .unbounded_send(message)
             .map(drop)
             .map_err(into_failure)
     }
+
     /// Broadcast transaction to other node.
     pub fn broadcast_transaction(&self, tx: Signed<RawTransaction>) -> Result<(), Error> {
         let msg = ExternalMessage::Transaction(tx);
@@ -845,8 +845,8 @@ pub struct NodeChannel {
     ),
     /// Channel for api requests.
     pub api_requests: (
-        mpsc::Sender<ExternalMessage>,
-        mpsc::Receiver<ExternalMessage>,
+        mpsc::UnboundedSender<ExternalMessage>,
+        mpsc::UnboundedReceiver<ExternalMessage>,
     ),
     /// Channel for network events.
     pub network_events: (mpsc::Sender<NetworkEvent>, mpsc::Receiver<NetworkEvent>),
@@ -871,7 +871,7 @@ impl NodeChannel {
         Self {
             network_requests: mpsc::channel(buffer_sizes.network_requests_capacity),
             internal_requests: mpsc::channel(buffer_sizes.internal_events_capacity),
-            api_requests: mpsc::channel(buffer_sizes.api_requests_capacity),
+            api_requests: mpsc::unbounded(), // TODO ECR-3163
             network_events: mpsc::channel(buffer_sizes.network_events_capacity),
             internal_events: mpsc::channel(buffer_sizes.internal_events_capacity),
         }